requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
fake-useragent==0.1.11
future==0.17.1
idna==2.8
isort==4.3.4
oauthlib==3.0.1
python-dateutil==2.8.1
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
soupsieve==1.8
toml==0.10.0
urllib3==1.24.2
python-telegram-bot==13.13